Rename link-button to text-button
