+++ packages/frontity-org-theme/src/processors/index.ts
@@ -4,7 +4,7 @@ export { default as textColor } from "./text-color";
 export { default as buttonDefault } from "./button-default";
 export { default as buttonBig } from "./button-big";
 export { default as buttonIcon } from "./button-icon";
-export { default as buttonLink } from "./button-link";
+export { default as buttonLink } from "./button-text";
 export { default as commonLink } from "./common-link";
 export { default as mobileDesktop } from "./mobile-desktop";
 export { default as specialIcons } from "./special-icons";
